# Conflicts:
#	src/middleware/auth.middleware.js

+++ src/controller/post.controller.js
@@ -37,7 +37,7 @@ exports.create = async (req, res) => {
     }
 }
 
-exports.getAll = async (req, res) => {
+exports.getAll = async (_req, res) => {
     const postList = await Post.find({}, {
         _id: 0,
         text: 1,
@@ -57,6 +57,10 @@ exports.update = async (req, res) => {
     const post = await Post.findOne({_id: postId});
     if (!post) {
         return res.status(404).json({error: "Le post n'existe pas"});
+    }
+
+    if (post.user_id.toString() !== req.token._id.toString()) {
+        return res.status(403).json({ error: "Vous ne pouvez pas modifier ce post" });
     }
 
     const newText = req.body.text;
@@ -86,6 +90,10 @@ exports.delete = async (req, res) => {
     const postId = req.params.id;
     if (!postId) {
         return res.status(400).json({error: "Vous devez spécifier un id de post"});
+    }
+
+    if (post.user_id.toString() !== req.token._id.toString()) {
+        return res.status(403).json({ error: "Vous ne pouvez pas supprimer ce post" });
     }
 
     const result = await Post.deleteOne({_id: postId});
